Refactorización de codigo innecesario
