+++ src/components/ui/table.tsx
@@ -1,5 +1,5 @@
 import type { FC } from "react";
-import { cn } from "@/utils/tailwind-cn";
+import { cn } from "@/lib/utils";
 
 interface Props {
   className?: string;
